pyproject.toml
[tool.poetry]
name = "dbbot-sqlalchemy"
version = "2.0-xam"
description = ""
authors = ["Your Name <you@example.com>"]
license = "Apache License Version 2"

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "1.4.31"
loguru = "0.6.0"
robotframework = "^4.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
